Minor cleanup of OpenTSDBWriterTests
This is all very minor, but makes the test easier to read IMHO. There is still a dependency on PowerMock. To keep the same test coverage, I should extract the resolving of local hostname, but that's more effort than I am willing to put into this at the moment. I almost think I should just remove those tests as they are really corner cases and are fairly costly.